+++ src/main/java/org/lagonette/hellos/bean/helloasso/CustomItemTypeEnum.java
@@ -1,0 +1,5 @@
+package org.lagonette.hellos.bean.helloasso;
+
+public enum CustomItemTypeEnum {
+    Date, TextInput, FreeText, ChoiceList, File, YesNo, Phone, Zipcode, Number
+}